Refactor Log class into RawLog and Log; Fix RawLog Constructor
The origial Log class was cumbersome to work with because it stored data
as the raw lists of lines from the log file. This required parsing those
lines whenever performing an operation. To make working with logs
easier, create a ``Log`` class and rename the former ``Log`` class to
``RawLog``. The ``Log`` class immediately parses the contents of the
log. As not all sections of the log have code to parse them, the ``Log``
class does not support all sections yet.

To allow making copies of ``RawLog`` objects, remove the ``log_file``
attribute and make the constructor initialize attributes to empty lists.
Then, those attributes can be set. This allows methods like
``from_file`` and ``from_raw_log`` to coexist.